Parallelize event indexing.
This commit refactors the indexing architecture. Previously,
we observed indexing rates in the order of 20-40k events/sec. The granularity
of concurrency was a single event. In comparison, the ingestion rate of a
bro conn.log having 5.4M lines was around 90-100k. While it took 80 seconds to
ingest the log, it was order 15 minutes to index it.

With this commit, we pushed down the granularity of indexing to one actor per
value. That is, for a conn.log with 20 fields/values, VAST uses 20 actors plus
two extra actors for event name and timestamp (22 actors total). Not
surprisingly, this changed the performance: it now takes 58 seconds to index
the events from the conn.log mentioned above, at a rate of also 90-100k
events/sec: indexing is now *faster* than ingestion, with a eman indexing rate
of 2.1M values/sec.

+++ src/vast/index.h
@@ -30,8 +30,8 @@ public:
   {
     struct partition_state
     {
-      size_t got = 0;
-      optional<size_t> expected;
+      uint64_t got = 0;
+      optional<uint64_t> expected;
       bitstream hits;
     };
 
@@ -64,7 +64,7 @@ public:
   /// @param pred The (existing) predicate to update.
   /// @param part The partition *pred* came from.
   /// @param n The number of expected hits for *pred* from *part*.
-  void expect(expr::ast const& pred, uuid const& part, size_t n);
+  void expect(expr::ast const& pred, uuid const& part, uint64_t n);
 
   /// Updates the state of a given partition.
   /// @param id The UUID of the partition to update.
